Added helpers as member functions and moved class to file_io, chnaged thread lock to fvcore file lock
